[duckdb] Relocated DuckDBDaVinciRecordTransformer to another package (#1441)
From c.l.v.sql to c.l.v.duckdb
